progress on refactoring on_primary_key

+++ src/db/sqlite/sqlite_statement.cpp
@@ -17,7 +17,7 @@ sqlite_statement::sqlite_statement(sqlite_connection &db, const matador::sql &sq
   , stmt_(nullptr)
 {
   // prepare sqlite statement
-  int ret = sqlite3_prepare_v2(db_.handle(), str().c_str(), str().size(), &stmt_, nullptr);
+  int ret = sqlite3_prepare_v2(db_.handle(), str().c_str(), static_cast<int>(str().size()), &stmt_, nullptr);
   throw_database_error(ret, db_.handle(), "sqlite3_prepare_v2", str());
   binder_ = std::make_unique<sqlite_parameter_binder>(db.handle(), stmt_);
 }
